requirements.txt

Django==1.9.5

djangorestframework==3.3.3

markdown==2.6.6

django-filter==0.13.0

django-mptt==0.8.3

pillow==3.2.0